I think I'm mostly done with the Movie module, I just need to test it and make sure the API is able to get hit to load the movie if it isn't already in the database.
Created Review module.

I added SQL statments to this script to create the movies table. I
decided to use the reviews table to model the many-to-many relationship
between users and movies. This puts the constraint of one review per
movie for each user, but will will add an edit featrue so the user can
change their review on a movie.

Still need to test, but that is tomorrow's problem.